rename scraper part of program
It is already known that this is the one and only scraper part of the app and it is related to EGS.
No need to explicitly state that in the title.
It has been dragged along since the earlier versions of the app and now it's time to fix this.